Feat: Add projects subscription to dashboard
